Add cheeps-per-user metric

+++ itu-minitwit/src/MiniTwit.Core/Services/IMetricsService.cs
@@ -4,5 +4,5 @@ public interface IMetricsService
 {
     void SetTotalUsers(int value);
     void SetTotalCheeps(int value);
-    void SetCheepsPerUsers(int value);
+    void SetCheepsPerUsers(float value);
 }
